feat(dashboard): re-design dashboard with tabbed navigation
