+++ bower.json
@@ -7,40 +7,39 @@
   ],
   "description": "Live, controllable graphics system rendered in a browser",
   "dependencies": {
-    "clipboard": "^1.5.9",
-    "cookies-js": "~1.2.1",
-    "draggabilly": "~1.2.4",
-    "iframe-resizer": "~3.4.2",
-    "imagesloaded": "~3.1.5",
-    "iron-collapse": "PolymerElements/iron-collapse#~1.0.5",
-    "iron-image": "PolymerElements/iron-image#^1.0.3",
-    "iron-localstorage": "PolymerElements/iron-localstorage#^1.0.4",
-    "iron-pages": "PolymerElements/iron-pages#^1.0.2",
-    "iron-selector": "PolymerElements/iron-selector#1.0.7",
-    "js-throttle-debounce": "~0.1.1",
-    "more-routing": "PolymerLabs/more-routing#^1.0.0",
+    "app-layout": "PolymerElements/app-layout#^0.10.7",
+    "app-route": "PolymerElements/app-route#^0.9.3",
+    "clipboard": "^1.6.1",
+    "cookies-js": "^1.2.3",
+    "draggabilly": "^2.1.1",
+    "iframe-resizer": "^3.5.11",
+    "imagesloaded": "^4.1.1",
+    "iron-collapse": "PolymerElements/iron-collapse#^1.3.0",
+    "iron-flex-layout": "PolymerElements/iron-flex-layout#^1.3.2",
+    "iron-icons": "PolymerElements/iron-icons#^1.2.0",
+    "iron-image": "PolymerElements/iron-image#^1.2.5",
+    "iron-localstorage": "PolymerElements/iron-localstorage#^1.0.6",
+    "iron-pages": "PolymerElements/iron-pages#^1.0.8",
+    "iron-selector": "PolymerElements/iron-selector#^1.5.2",
+    "js-throttle-debounce": "^0.1.1",
     "nodecg-replicant": "NodeCGElements/nodecg-replicant#^0.5.8",
-    "observe-shim": "~0.4.2",
-    "packery": "~1.4.3",
-    "paper-button": "PolymerElements/paper-button#^1.0.10",
-    "paper-card": "PolymerElements/paper-card#^1.0.8",
-    "paper-dialog": "PolymerElements/paper-dialog#^1.0.3",
-    "paper-dialog-behavior": "PolymerElements/paper-dialog-behavior#^1.1.0",
-    "paper-dialog-scrollable": "PolymerElements/paper-dialog-scrollable#^1.1.0",
-    "paper-drawer-panel": "PolymerElements/paper-drawer-panel#^1.0.5",
-    "paper-header-panel": "PolymerElements/paper-header-panel#^1.1.2",
-    "paper-icon-button": "PolymerElements/paper-icon-button#^1.0.5",
-    "paper-item": "PolymerElements/paper-item#^1.1.4",
-    "paper-slider": "PolymerElements/paper-slider#^1.0.10",
-    "paper-spinner": "PolymerElements/paper-spinner#^1.1.1",
-    "paper-styles": "PolymerElements/paper-styles#^1.0.13",
-    "paper-toast": "PolymerElements/paper-toast#^1.1.1",
-    "paper-toolbar": "PolymerElements/paper-toolbar#^1.1.2",
-    "rollbar": "^1.8.5",
-    "SoundJS": "0.6.2",
-    "vaadin-upload": "vaadin/vaadin-upload#^1.0.0-alpha1",
-    "iron-icons": "PolymerElements/iron-icons#^1.1.3",
-    "iron-flex-layout": "PolymerElements/iron-flex-layout#^1.3.2"
+    "packery": "^2.1.1",
+    "paper-button": "PolymerElements/paper-button#^1.0.14",
+    "paper-card": "PolymerElements/paper-card#^1.1.4",
+    "paper-dialog": "PolymerElements/paper-dialog#^1.1.0",
+    "paper-dialog-behavior": "PolymerElements/paper-dialog-behavior#^1.2.7",
+    "paper-dialog-scrollable": "PolymerElements/paper-dialog-scrollable#^1.1.5",
+    "paper-icon-button": "PolymerElements/paper-icon-button#^1.1.4",
+    "paper-item": "PolymerElements/paper-item#^1.2.1",
+    "paper-slider": "PolymerElements/paper-slider#^1.0.14",
+    "paper-spinner": "PolymerElements/paper-spinner#^1.2.1",
+    "paper-styles": "PolymerElements/paper-styles#^1.2.0",
+    "paper-tabs": "PolymerElements/paper-tabs#^1.8.0",
+    "paper-toast": "PolymerElements/paper-toast#^1.3.0",
+    "paper-toolbar": "PolymerElements/paper-toolbar#^1.1.7",
+    "rollbar": "^1.9.4",
+    "SoundJS": "^0.6.2",
+    "vaadin-upload": "vaadin/vaadin-upload#^1.1.3"
   },
   "license": "MIT",
   "private": true,
