+++ lib/dashboard/index.js
@@ -1,30 +1,34 @@
 'use strict';
 
+// Native
 const fs = require('fs');
+const path = require('path');
+const pug = require('pug');
+
+// Packages
 const express = require('express');
+const clone = require('clone');
+
+// Ours
 const configHelper = require('../config');
 const log = require('../logger')('nodecg/lib/dashboard');
 const bundles = require('../bundles');
-const path = require('path');
 const ncgUtils = require('../util');
-const jade = require('jade');
+
 const app = express();
-const publicPath = path.join(__dirname, 'public');
+const PUBLIC_PATH = path.join(__dirname, 'public');
 let cachedDashboard = null;
 
 log.trace('Adding Express routes');
-app.use('/dashboard', express.static(publicPath));
-app.set('views', publicPath);
+app.use('/bower_components', express.static(path.resolve(__dirname, '../../bower_components')));
+app.use('/dashboard', express.static(PUBLIC_PATH));
+app.set('views', PUBLIC_PATH);
 
 app.get('/', (req, res) => res.redirect('/dashboard/'));
 
 app.get('/dashboard', ncgUtils.authCheck, (req, res) => {
 	if (configHelper.config.developer) {
-		res.render(path.join(__dirname, 'src/dashboard.jade'), {
-			bundles: bundles.all(),
-			publicConfig: configHelper.filteredConfig,
-			privateConfig: configHelper.config
-		});
+		res.send(renderDashboard());
 	} else {
 		res.status(200).send(cachedDashboard);
 	}
@@ -94,14 +98,6 @@ app.get('/panel/:bundle/*', ncgUtils.authCheck, (req, res, next) => {
 	}
 });
 
-function cacheDashboard() {
-	cachedDashboard = jade.renderFile(path.join(__dirname, 'src/dashboard.jade'), {
-		bundles: bundles.all(),
-		publicConfig: configHelper.filteredConfig,
-		privateConfig: configHelper.config
-	});
-}
-
 // Initialize cache
 cacheDashboard();
 
@@ -109,3 +105,29 @@ cacheDashboard();
 bundles.on('bundleChanged', cacheDashboard);
 
 module.exports = app;
+
+function renderDashboard() {
+	return pug.renderFile(path.join(__dirname, 'src/dashboard.pug'), {
+		bundles: bundles.all().map(bundle => {
+			const cleanedBundle = clone(bundle);
+			if (cleanedBundle.dashboardPanels) {
+				cleanedBundle.dashboardPanels = cleanedBundle.dashboardPanels.forEach(panel => {
+					delete panel.html;
+				});
+			}
+
+			delete cleanedBundle.rawManifest;
+			return cleanedBundle;
+		}),
+		publicConfig: configHelper.filteredConfig,
+		privateConfig: configHelper.config,
+		workspaces: [{
+			name: 'Primary Workspace',
+			route: ''
+		}]
+	});
+}
+
+function cacheDashboard() {
+	cachedDashboard = renderDashboard();
+}
