+++ lib/dashboard/public/dashboard.css
@@ -1,69 +1,5 @@
-body {
-  background-color: #dedede;
-}
-#panels {
-  width: 100%;
-  height: 100%;
-  padding-bottom: 32px;
-  transition: opacity 500ms ease, transform 600ms ease-out;
-  opacity: 0;
-  pointer-events: none;
-  transform: translateY(40px);
-}
-@media (max-width: 640px) {
-  #panels {
-    padding: 0 6px;
-  }
-}
-.dashboard-panel {
-  width: 128px;
-}
-.dashboard-panel.width-1 {
-  width: 128px;
-}
-.dashboard-panel.width-2 {
-  width: 272px;
-}
-.dashboard-panel.width-3 {
-  width: 416px;
-}
-.dashboard-panel.width-4 {
-  width: 560px;
-}
-.dashboard-panel.width-5 {
-  width: 704px;
-}
-.dashboard-panel.width-6 {
-  width: 848px;
-}
-.dashboard-panel.width-7 {
-  width: 992px;
-}
-.dashboard-panel.width-8 {
-  width: 1136px;
-}
-.logo {
-  vertical-align: middle;
-  margin-right: 1em;
-}
 .text-warning {
   color: #c53929;
-}
-.panel-heading {
-  background-color: #616161;
-  color: #f5f5f5;
-  opacity: 1;
-  padding: 10px 15px;
-}
-.panel-heading.green {
-  background-color: #0b8043;
-}
-.panel-body {
-  color: #616161;
-  padding: 15px;
-}
-.navbar {
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.4);
 }
 #ncg-main-toolbar {
   color: #f5f5f5;
@@ -95,7 +31,7 @@ body {
 #ncg-pages section {
   padding: 32px;
 }
-@media (max-width: 640px) {
+@media (max-width: var(--width-sm)) {
   #ncg-pages section {
     padding: 12px 0;
   }
@@ -113,27 +49,6 @@ body {
 }
 ncg-dashboard-panel iframe {
   vertical-align: middle;
-}
-#instancesList-empty {
-  text-align: center;
-  border: 4px dashed #B7B7B7;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 24px;
-}
-.graphic {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.graphic:not(:last-child) {
-  margin-bottom: 6px;
-}
-.graphic paper-button {
-  white-space: nowrap;
-  flex-shrink: 0;
 }
 .assets-divider {
   box-sizing: border-box;
