+++ lib/dashboard/public/elements/assets/ncg-asset-file.js
