+++ lib/dashboard/public/elements/assets/ncg-asset-group.html
@@ -1,0 +1,101 @@
+<link rel="import" href="/bower_components/iron-flex-layout/iron-flex-layout.html">
+<link rel="import" href="/bower_components/iron-icons/iron-icons.html">
+<link rel="import" href="/bower_components/nodecg-replicant/nodecg-replicant.html">
+<link rel="import" href="/bower_components/paper-card/paper-card.html">
+<link rel="import" href="/bower_components/paper-toast/paper-toast.html">
+<link rel="import" href="/bower_components/polymer/polymer.html">
+<link rel="import" href="/bower_components/vaadin-upload/vaadin-upload.html">
+<link rel="import" href="ncg-asset-file.html">
+
+<dom-module id="ncg-asset-group">
+	<template>
+		<style>
+			:host {
+				display: block;
+				width: 100%;
+				box-sizing: border-box;
+			}
+
+			#add {
+				display: flex;
+				align-items: center;
+			}
+
+			paper-card {
+				width: 100%;
+			}
+
+			#header {
+				@apply(--layout-horizontal);
+				@apply(--layout-justified);
+				margin-bottom: 6px;
+			}
+
+			#title {
+				@apply(--paper-font-headline);
+			}
+
+			#files {
+				@apply(--layout-vertical);
+			}
+
+			vaadin-upload {
+				width: 400px;
+				margin: 16px;
+			}
+
+			#acceptsMsg {
+				margin-top: 8px;
+				text-align: center;
+			}
+		</style>
+
+		<nodecg-replicant
+			id="replicant"
+			replicant-name="assets:[[categoryName]]"
+			replicant-bundle="[[bundleName]]"
+		  	value="{{files}}"></nodecg-replicant>
+		<paper-toast id="toast"></paper-toast>
+
+		<div id="header">
+			<span id="title">[[categoryTitle]]</span>
+			<paper-button id="add" raised class="nodecg-configure" on-click="openUploadDialog">
+				<iron-icon icon="add"></iron-icon>
+				Add File(s)
+			</paper-button>
+		</div>
+
+		<div id="files">
+			<template is="dom-repeat" items="[[files]]" as="file">
+				<ncg-asset-file
+					file="[[file]]"
+				   	on-deleted="_handleDeleted"
+				   	on-deletion-failed="_handleDeletionFailed"></ncg-asset-file>
+			</template>
+		</div>
+
+		<div id="empty">
+			There are no assets in this category.
+		</div>
+
+		<!-- 2017/03/18: Had to remove with-backdrop during the dashboard re-write -->
+		<paper-dialog id="uploadDialog">
+			<vaadin-upload
+				id="uploader"
+				target="/assets/[[bundleName]]/[[categoryName]]"
+			   	on-upload-before="_onUploadBefore"
+				on-file-reject="_onFileReject">
+
+				<template is="dom-if" if="[[allowedTypes.length]]">
+					<div id="acceptsMsg">[[acceptsMsg]]</div>
+				</template>
+			</vaadin-upload>
+
+			<div class="buttons">
+				<paper-button dialog-dismiss>Close</paper-button>
+			</div>
+		</paper-dialog>
+	</template>
+
+	<script src="ncg-asset-group.js"></script>
+</dom-module>
