+++ lib/dashboard/public/elements/assets/ncg-asset-group.js
@@ -2,7 +2,7 @@
 	'use strict';
 
 	Polymer({
-		is: 'ncg-assets',
+		is: 'ncg-asset-group',
 
 		properties: {
 			files: Array,
