+++ lib/dashboard/public/elements/assets/ncg-assets.html
@@ -1,0 +1,52 @@
+<link rel="import" href="/bower_components/paper-card/paper-card.html">
+<link rel="import" href="ncg-asset-group.html">
+
+<dom-module id="ncg-assets">
+	<template>
+		<style>
+			:host {
+				width: 100%;
+				@apply(--layout-vertical);
+				@apply(--layout-center);
+			}
+
+			.assets-divider {
+				box-sizing: border-box;
+				margin: 12px 10px;
+				border-top: 1px solid #bdbdbd;
+			}
+
+			.assets-divider:last-child {
+				display: none;
+			}
+		</style>
+
+		<template is="dom-repeat" items="[[bundlesWithAssets]]" as="bundle">
+			<paper-card class="ncg-card">
+				<div class="card-content">
+					<h2>[[bundle.name]]</h2>
+
+					<template is="dom-if" if="[[bundle.hasAssignableSoundCues]]">
+						<ncg-asset-group
+							bundle-name="[[bundle.name]]"
+							category-name="sounds"
+							category-title="Sounds"
+							allowed-types='["mp3","ogg"]'></ncg-asset-group>
+						<div class="assets-divider"></div>
+					</template>
+
+					<template is="dom-repeat" items="[[bundle.assetCategories]]" as="category">
+						<ncg-asset-group
+							bundle-name="[[bundle.name]]"
+							category-name="[[category.name]]"
+							category-title="[[category.title]]"
+							allowed-types="[[category.allowedTypes]]"></ncg-asset-group>
+						<div class="assets-divider"></div>
+					</template>
+				</div>
+			</paper-card>
+		</template>
+	</template>
+
+	<script src="ncg-assets.js"></script>
+</dom-module>
