+++ lib/dashboard/public/elements/graphics/ncg-single-instance.html
@@ -1,0 +1,19 @@
+<link rel="import" href="/bower_components/paper-button/paper-button.html">
+
+<dom-module id="ncg-single-instance">
+	<template>
+		<style>
+			:host {
+				display: flex;
+				align-items: center;
+				justify-content: space-between;
+				margin-bottom: 8px;
+			}
+		</style>
+
+		<span>{{url}}</span>
+		<paper-button class="nodecg-danger" on-click="kill">KILL</paper-button>
+	</template>
+
+	<script src="ncg-single-instance.js"></script>
+</dom-module>
