+++ lib/dashboard/public/elements/graphics/ncg-single-instance.js
