+++ lib/dashboard/public/elements/mixer/ncg-sound-cue.html
@@ -1,8 +1,8 @@
-<link rel="import" href="/components/polymer/polymer.html">
-<link rel="import" href="/components/iron-flex-layout/iron-flex-layout.html">
-<link rel="import" href="/components/iron-flex-layout/iron-flex-layout-classes.html">
-<link rel="import" href="/components/paper-slider/paper-slider.html">
-<link rel="import" href="/components/paper-styles/typography.html">
+<link rel="import" href="/bower_components/polymer/polymer.html">
+<link rel="import" href="/bower_components/iron-flex-layout/iron-flex-layout.html">
+<link rel="import" href="/bower_components/iron-flex-layout/iron-flex-layout-classes.html">
+<link rel="import" href="/bower_components/paper-slider/paper-slider.html">
+<link rel="import" href="/bower_components/paper-styles/typography.html">
 
 <dom-module id="ncg-sound-cue">
 	<template>
