+++ lib/dashboard/public/elements/mixer/ncg-sound-cue.js
