+++ lib/dashboard/public/elements/mixer/ncg-sounds.html
@@ -1,9 +1,8 @@
-<link rel="import" href="/components/polymer/polymer.html">
-<link rel="import" href="/components/paper-card/paper-card.html">
-<link rel="import" href="/components/paper-dialog/paper-dialog.html">
-<link rel="import" href="/components/paper-dialog-scrollable/paper-dialog-scrollable.html">
-<link rel="import" href="/components/paper-toast/paper-toast.html">
-<link rel="import" href="/dashboard/elements/ncg-sound-cue.html">
+<link rel="import" href="/bower_components/paper-card/paper-card.html">
+<link rel="import" href="/bower_components/paper-dialog/paper-dialog.html">
+<link rel="import" href="/bower_components/paper-dialog-scrollable/paper-dialog-scrollable.html">
+<link rel="import" href="/bower_components/paper-toast/paper-toast.html">
+<link rel="import" href="ncg-sound-cue.html">
 
 <dom-module id="ncg-sounds">
 	<template>
