--- lib/dashboard/public/elements/ncg-asset-file.html
@@ -1,43 +1,0 @@
-<link rel="import" href="/components/iron-flex-layout/iron-flex-layout.html">
-<link rel="import" href="/components/iron-icons/iron-icons.html">
-<link rel="import" href="/components/paper-button/paper-button.html">
-<link rel="import" href="/components/paper-spinner/paper-spinner-lite.html">
-<link rel="import" href="/components/polymer/polymer.html">
-
-<dom-module id="ncg-asset-file">
-	<style>
-		:host {
-			@apply(--layout-horizontal);
-			@apply(--layout-justified);
-			@apply(--layout-center);
-		}
-
-		#name {
-			text-transform: none;
-			overflow: hidden;
-			text-overflow: ellipsis;
-		}
-
-		#delete {
-			display: flex;
-			align-items: center;
-			color: var(--nodecg-danger-color);
-			flex-shrink: 0;
-		}
-
-		#spinner {
-			--paper-spinner-color: var(--nodecg-danger-color);
-		}
-	</style>
-
-	<template>
-		<a id="name" href="[[file.url]]" target="_blank">[[file.base]]</a>
-		<paper-spinner-lite id="spinner" alt="Deleting" active></paper-spinner-lite>
-		<paper-button id="delete" on-click="_handleDeleteClick">
-			<iron-icon icon="delete"></iron-icon>
-			Delete
-		</paper-button>
-	</template>
-</dom-module>
-
-<script src="/dashboard/elements/ncg-asset-file.js"></script>
