+++ lib/dashboard/public/elements/ncg-dashboard-panel.html
@@ -1,80 +1,113 @@
-<link rel="import" href="/components/polymer/polymer.html">
-<link rel="import" href="/components/paper-material/paper-material.html">
-<link rel="import" href="/components/paper-styles/paper-styles.html">
-<link rel="import" href="/components/paper-styles/paper-styles-classes.html">
-<link rel="import" href="/components/iron-collapse/iron-collapse.html">
-<link rel="import" href="/components/iron-flex-layout/iron-flex-layout.html">
-<link rel="import" href="/components/iron-localstorage/iron-localstorage.html">
+<link rel="import" href="/bower_components/paper-material/paper-material.html">
+<link rel="import" href="/bower_components/paper-styles/typography.html">
+<link rel="import" href="/bower_components/iron-collapse/iron-collapse.html">
+<link rel="import" href="/bower_components/iron-flex-layout/iron-flex-layout.html">
+<link rel="import" href="/bower_components/iron-localstorage/iron-localstorage.html">
 <link rel="import" href="ncg-dashboard-panel-info-dialog.html">
 
 <dom-module id="ncg-dashboard-panel">
-	<style>
-		:host {
-			display: inline-block;
-		}
-
-		#header {
-			color: var(--google-grey-100);
-			@apply(--layout-horizontal);
-			@apply(--layout-end-justified);
-			@apply(--layout-center);
-			overflow: hidden;
-		}
-
-		#displayTitle {
-			position: absolute;
-			left: 15px;
-			top: 6px;
-		}
-
-		#buttons {
-			@apply(--layout-horizontal);
-			@apply(--layout-center);
-			padding-left: 8px;
-			transform: translateX(100%);
-			transition: transform 200ms ease;
-		}
-
-		#more {
-			position: absolute;
-			right: 10px;
-			top: 8px;
-		}
-
-		#buttonsContainer {
-			z-index: 1;
-		}
-
-		#header:hover #buttons {
-			transform: translateX(0%);
-		}
-
-		#dragHandle {
-			cursor: -webkit-grab;
-			cursor: -moz-grab;
-			cursor: grab;
-		}
-
-		#dragHandle:active {
-			cursor: -webkit-grabbing;
-			cursor: -moz-grabbing;
-			cursor: grabbing;
-		}
-
-		#body {
-			padding: 0;
-			background-color: #f5f5f5;
-		}
-	</style>
 	<template>
-		<iron-localstorage name="{{computeLocalStorageName(bundle, panel)}}"
-						   value="{{opened}}"
-						   on-iron-localstorage-load-empty="initializeDefaultOpened"
-		></iron-localstorage>
+		<style>
+			:host {
+				display: inline-block;
+				width: 128px;
+			}
+
+			:host([width="1"]) {
+				width: 128px;
+			}
+
+			:host([width="2"]) {
+				width: 272px;
+			}
+
+			:host([width="3"]) {
+				width: 416px;
+			}
+
+			:host([width="4"]) {
+				width: 560px;
+			}
+
+			:host([width="5"]) {
+				width: 704px;
+			}
+
+			:host([width="6"]) {
+				width: 848px;
+			}
+
+			:host([width="7"]) {
+				width: 992px;
+			}
+
+			:host([width="8"]) {
+				width: 1136px;
+			}
+
+			#header {
+				color: var(--google-grey-100);
+				@apply(--layout-horizontal);
+				@apply(--layout-end-justified);
+				@apply(--layout-center);
+				@apply(--paper-font-title);
+				overflow: hidden;
+			}
+
+			#displayTitle {
+				position: absolute;
+				left: 15px;
+				top: 6px;
+			}
+
+			#buttons {
+				@apply(--layout-horizontal);
+				@apply(--layout-center);
+				padding-left: 8px;
+				transform: translateX(100%);
+				transition: transform 200ms ease;
+			}
+
+			#more {
+				position: absolute;
+				right: 10px;
+				top: 8px;
+			}
+
+			#buttonsContainer {
+				z-index: 1;
+			}
+
+			#header:hover #buttons {
+				transform: translateX(0%);
+			}
+
+			#dragHandle {
+				cursor: -webkit-grab;
+				cursor: -moz-grab;
+				cursor: grab;
+			}
+
+			#dragHandle:active {
+				cursor: -webkit-grabbing;
+				cursor: -moz-grabbing;
+				cursor: grabbing;
+			}
+
+			#body {
+				padding: 0;
+				background-color: #f5f5f5;
+			}
+		</style>
+
+		<iron-localstorage
+			name="{{computeLocalStorageName(bundle, panel)}}"
+		   	value="{{opened}}"
+		   	on-iron-localstorage-load-empty="initializeDefaultOpened"></iron-localstorage>
 
 		<paper-material>
 			<div id="header">
-				<span id="displayTitle" class="paper-font-title">[[displayTitle]]</span>
+				<span id="displayTitle">[[displayTitle]]</span>
 				<div id="buttonsContainer">
 					<iron-icon id="more" icon="chevron-left"></iron-icon>
 					<div id="buttons">
@@ -91,6 +124,7 @@
 			</iron-collapse>
 		</paper-material>
 	</template>
+
+	<script src="/dashboard/elements/ncg-dashboard-panel.js"></script>
 </dom-module>
 
-<script src="/dashboard/elements/ncg-dashboard-panel.js"></script>
