+++ lib/dashboard/public/elements/ncg-dashboard-panel.js
@@ -25,6 +25,10 @@
 				type: String,
 				reflectToAttribute: true,
 				observer: '_headerColorChanged'
+			},
+			width: {
+				type: Number,
+				reflectToAttribute: true
 			}
 		},
