+++ lib/dashboard/public/elements/ncg-dialog.html
@@ -1,11 +1,11 @@
-<link rel="import" href="/components/polymer/polymer.html">
-<link rel="import" href="/components/neon-animation/neon-animation-runner-behavior.html">
-<link rel="import" href="/components/paper-dialog-behavior/paper-dialog-behavior.html">
-<link rel="import" href="/components/paper-dialog-behavior/paper-dialog-shared-styles.html">
+<link rel="import" href="/bower_components/polymer/polymer.html">
+<link rel="import" href="/bower_components/neon-animation/neon-animation-runner-behavior.html">
+<link rel="import" href="/bower_components/paper-dialog-behavior/paper-dialog-behavior.html">
+<link rel="import" href="/bower_components/paper-dialog-behavior/paper-dialog-shared-styles.html">
 
 <!--
 You may notice some oddness with this element. Namely that the `.buttons` div appears before the
-`paper-dialog-scrollable` element in `dashboard.jade`, but in the below `<style>` tag I've forced
+`paper-dialog-scrollable` element in `dashboard.pug`, but in the below `<style>` tag I've forced
 `.buttons` to have an `order` of `3`, making it appear after `paper-dialog-scrollable`.
 
 This was done to fix a bizarre issue wherein any nodes placed after the `paper-dialog-scrollable`
@@ -16,43 +16,38 @@ Lange - 1/28/2016
 -->
 
 <dom-module id="ncg-dialog">
-	<style>
-		:host {
-			flex-direction: column;
-			max-width: 100%;
-			--paper-dialog-scrollable: {
-				background-color: wheat;
-				max-width: none;
-			}
-		}
-
-		:host > ::content iframe {
-			box-sizing: border-box;
-		}
-
-		:host > ::content iframe {
-			margin: 0 !important;
-			padding: 0 !important;
-		}
-
-		:host > ::content .buttons {
-			order: 3;
-		}
-	</style>
-
 	<template>
-		<style include="paper-dialog-shared-styles"></style>
-		<style>
+		<style include="paper-dialog-shared-styles">
 			:host {
 				display: flex;
+				flex-direction: column;
+				max-width: 100%;
+				--paper-dialog-scrollable: {
+					background-color: wheat;
+					max-width: none;
+				}
+			}
+
+			:host > ::content iframe {
+				box-sizing: border-box;
+			}
+
+			:host > ::content iframe {
+				margin: 0 !important;
+				padding: 0 !important;
+			}
+
+			:host > ::content .buttons {
+				order: 3;
 			}
 
 			:host > ::content > *:last-child {
 				margin-bottom: 0;
 			}
 		</style>
+
 		<content></content>
 	</template>
-</dom-module>
 
-<script src="/dashboard/elements/ncg-dialog.js"></script>
+	<script src="/dashboard/elements/ncg-dialog.js"></script>
+</dom-module>
