--- lib/dashboard/public/elements/ncg-single-instance.html
@@ -1,21 +1,0 @@
-<link rel="import" href="/components/polymer/polymer.html">
-<link rel="import" href="/components/paper-styles/paper-styles.html">
-<link rel="import" href="/components/paper-button/paper-button.html">
-
-<dom-module id="ncg-single-instance">
-	<style>
-		:host {
-			display: flex;
-			align-items: center;
-			justify-content: space-between;
-			margin-bottom: 8px;
-		}
-	</style>
-
-	<template>
-		<span>{{url}}</span>
-		<paper-button class="nodecg-danger" on-click="kill">KILL</paper-button>
-	</template>
-</dom-module>
-
-<script src="/dashboard/elements/ncg-single-instance.js"></script>
