+++ lib/dashboard/public/elements/settings/ncg-settings.html
@@ -1,0 +1,86 @@
+<link rel="import" href="/bower_components/iron-flex-layout/iron-flex-layout.html">
+<link rel="import" href="/bower_components/iron-icons/iron-icons.html">
+<link rel="import" href="/bower_components/paper-button/paper-button.html">
+<link rel="import" href="/bower_components/paper-card/paper-card.html">
+<link rel="import" href="/bower_components/paper-dialog/paper-dialog.html">
+<link rel="import" href="/bower_components/paper-toast/paper-toast.html">
+
+<dom-module id="ncg-settings">
+	<template>
+		<style>
+			:host {
+				width: 100%;
+				@apply(--layout-vertical);
+				@apply(--layout-center);
+			}
+
+			#copyKey {
+				color: white;
+				background-color: #00bebe;
+			}
+
+			.card-actions {
+				@apply(--layout-horizontal);
+				@apply(--layout-center-center);
+			}
+
+			paper-button {
+				@apply(--layout-horizontal);
+				@apply(--layout-center-center);
+			}
+		</style>
+
+		<paper-card heading="Your Key" class="ncg-card">
+			<div class="card-content">
+				<p class="text-warning paper-font-body1">
+					Resetting your key will disrupt all current sessions using it.<br/>
+					When you reset your key, the dashboard will be refreshed so that a new key can be obtained.
+				</p>
+				<div class="card-actions">
+					<paper-button raised  id="copyKey">
+						<iron-icon icon="icons:content-copy"></iron-icon>&ensp;Copy Key
+					</paper-button>
+					<paper-button raised id="showKey" class="nodecg-configure" title="Show Key" on-tap="openShowKeyDialog">
+						<iron-icon icon="communication:vpn-key"></iron-icon>&ensp;Show Key
+					</paper-button>
+					<paper-button raised id="resetKey" class="nodecg-danger" on-tap="openResetKeyDialog">
+						<iron-icon icon="icons:refresh"></iron-icon>&ensp;Reset Key
+					</paper-button>
+				</div>
+			</div>
+		</paper-card>
+
+		<!-- 2017/03/18: Had to remove with-backdrop during the dashboard re-write -->
+		<paper-dialog id="showKeyDialog">
+			<h2>NodeCG Key</h2>
+			<div>
+				<code id="key" class="paper-font-code1"></code>
+				<p class="text-warning paper-font-body1">
+					<b>Do not</b> give this key to anyone or show it on stream!<br/>
+					If you accidentally reveal it, <b>reset it immediately!</b>
+				</p>
+			</div>
+			<div class="buttons">
+				<paper-button dialog-dismiss="dialog-dismiss">Close</paper-button>
+			</div>
+		</paper-dialog>
+
+		<paper-dialog id="resetKeyDialog">
+			<h2>Reset NodeCG Key</h2>
+			<div>
+				<p class="text-warning paper-font-body1">
+					Are you sure you wish to reset your <b>NodeCG key</b>?<br/>
+					Doing so will invalidate all URLs currently loaded into your streaming software!
+				</p>
+			</div>
+			<div class="buttons">
+				<paper-button dialog-dismiss="dialog-dismiss">No, Cancel</paper-button>
+				<paper-button dialog-dismiss="dialog-confirm" on-tap="resetKey">Yes, reset</paper-button>
+			</div>
+		</paper-dialog>
+
+		<paper-toast id="settingsToast"></paper-toast>
+	</template>
+
+	<script src="ncg-settings.js"></script>
+</dom-module>
