--- lib/dashboard/src/assets.jade
@@ -1,14 +1,0 @@
-each bundle in bundles
-	if (bundle.hasAssignableSoundCues || (bundle.assetCategories && bundle.assetCategories.length > 0))
-		paper-card.ncg-card
-			.card-content
-				h2=bundle.name
-				if (bundle.hasAssignableSoundCues)
-					ncg-assets.assets-category(bundle-name=bundle.name, category-name='sounds',
-					category-title="Sounds", allowed-types="[\"mp3\",\"ogg\"]")
-					.assets-divider
-				each category in bundle.assetCategories
-					- var strinigifiedTypes = JSON.stringify(category.allowedTypes)
-					ncg-assets.assets-category(bundle-name=bundle.name, category-name=category.name,
-					category-title=category.title, allowed-types=strinigifiedTypes)
-					.assets-divider
