--- lib/dashboard/src/assets.less
@@ -1,9 +1,0 @@
-.assets-divider {
-  &:last-child {
-	display: none;
-  }
-
-  box-sizing: border-box;
-  margin: 12px 10px;
-  border-top: 1px solid #bdbdbd;
-}
