--- lib/dashboard/src/graphics.jade
@@ -1,17 +1,0 @@
-paper-card.ncg-card(heading='Active Single-Instance Graphics')
-	.card-content
-		#instancesList
-		#instancesList-empty
-			span No single-instance graphics are currently open.
-each bundle in bundles
-	if (bundle.graphics.length > 0)
-		paper-card.ncg-card(heading=bundle.name)
-			.card-content.layout.vertical
-				each graphic in bundle.graphics
-					.graphic
-						div(style='display: flex; flex-direction: column;')
-							a(href=graphic.url, target='_blank')=graphic.url
-							span #{graphic.width}x#{graphic.height}
-						paper-button.js-copy(data-copy-type='graphic')
-							iron-icon(icon='icons:content-copy')
-							| &ensp;Copy URL
