--- lib/dashboard/src/graphics.less
@@ -1,24 +1,0 @@
-#instancesList-empty {
-  text-align: center;
-  border: 4px dashed #B7B7B7;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 24px;
-}
-
-.graphic {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  &:not(:last-child) {
-	margin-bottom: 6px;
-  }
-}
-
-.graphic paper-button {
-	white-space: nowrap;
-	flex-shrink: 0;
-}
