--- lib/dashboard/src/settings.jade
@@ -1,26 +1,0 @@
-if privateConfig.login.enabled
-	paper-card.ncg-card(heading='Your Key')
-		.card-content
-			p.text-warning.paper-font-body1
-				| Resetting your key will disrupt all current sessions using it.<br/>
-				| When you reset your key, the dashboard will be refreshed so that a new key can be obtained.
-			.card-actions
-				paper-button#showKey(alt='Show Key')
-					iron-icon(icon='communication:vpn-key')
-					| &ensp;Show Key
-				paper-button#copyKey.js-copy
-					iron-icon(icon='icons:content-copy')
-					| &ensp;Copy Key
-				paper-button#resetKey
-					iron-icon(icon='icons:refresh')
-					| &ensp;Reset Key
-
-	paper-dialog#showKeyDialog(with-backdrop)
-		h2 NodeCG Key
-		div
-			code.paper-font-code1#key
-			p.text-warning.paper-font-body1
-				| Do not give this key to anyone or show it on stream!<br/>
-				| If you accidentally reveal it, reset it immediately.
-		.buttons
-			paper-button(dialog-dismiss) Close
