+++ lib/login/index.js
@@ -85,14 +85,14 @@ app.use('/login', express.static(path.join(__dirname, 'public')));
 app.set('views', __dirname);
 
 app.get('/login', (req, res) => {
-	res.render('public/login.jade', {
+	res.render('public/login.pug', {
 		user: req.user,
 		config
 	});
 });
 
 app.get('/authError', (req, res) => {
-	res.render('public/authError.jade', {
+	res.render('public/authError.pug', {
 		message: req.query.message,
 		code: req.query.code,
 		viewUrl: req.query.viewUrl
