+++ lib/login/public/authError.pug
@@ -5,8 +5,8 @@ html(lang="en")
         meta(name="viewport" content="width=device-width" initial-scale=1)
         title Authorization Error
         //Bootstrap main CSS & Yeti override
-        link(rel='stylesheet' href='/components/bootstrap/dist/css/bootstrap.min.css')
-        link(rel='stylesheet' href='/components/bsw-yeti/css/bootstrap.min.css')
+        link(rel='stylesheet' href='/bower_components/bootstrap/dist/css/bootstrap.min.css')
+        link(rel='stylesheet' href='/bower_components/bsw-yeti/css/bootstrap.min.css')
         link(rel='stylesheet' href='/login/authError.css')
     body
         div.container
