+++ lib/login/public/login.pug
@@ -6,7 +6,7 @@ html(lang="en")
 
 		title NodeCG login
 
-		link(rel='import', href='/components/paper-styles/typography.html')
+		link(rel='import', href='/bower_components/paper-styles/typography.html')
 
 		style(is='custom-style' include='iron-flex iron-flex-alignment').
 			body {
