+++ lib/server/index.js
@@ -95,11 +95,10 @@ module.exports.start = function () {
 	// Set up Express
 	log.trace('Setting up Express');
 	app.use(require('compression')());
-	app.use('/components', express.static(path.resolve(__dirname, '../../bower_components')));
 	app.use(bodyParser.json());
 	app.use(bodyParser.urlencoded({extended: true}));
 
-	app.engine('jade', require('jade').__express);
+	app.engine('pug', require('pug').__express);
 
 	if (config.login && config.login.enabled) {
 		log.info('Login security enabled');
