+++ lib/util/injectscripts.js
@@ -55,7 +55,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 
 			if (opts.standalone) {
 				// Load the socket
-				scripts.push('<script src="/components/cookies-js/dist/cookies.min.js"></script>');
+				scripts.push('<script src="/bower_components/cookies-js/dist/cookies.min.js"></script>');
 				scripts.push('<script src="/socket.io/socket.io.js"></script>');
 
 				scripts.push(`<script>
@@ -74,7 +74,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 
 			// Likewise, they both need the contentWindow portion of the iframeResizer.
 			// We put this at the start and make it async so it loads ASAP.
-			scripts.unshift('<script async src="/components/iframe-resizer/js/iframeResizer.contentWindow.min.js">' +
+			scripts.unshift('<script async src="/bower_components/iframe-resizer/js/iframeResizer.contentWindow.min.js">' +
 				'</script>');
 
 			// Panels need the default panel styles and the dialog_opener.
@@ -88,7 +88,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 		} else if (resourceType === 'graphic') {
 			if (config.rollbar && config.rollbar.enabled) {
 				scripts.unshift(
-					'<script src="/components/rollbar/dist/rollbar.min.js"></script>',
+					'<script src="/bower_components/rollbar/dist/rollbar.min.js"></script>',
 					`<script>
 						Rollbar.init({
 							accessToken: '${config.rollbar.postClientItem}',
@@ -108,7 +108,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 			// If login security is enabled, graphics must provide a valid token when they connect to the socket
 			// They also need cookies-js and a helper to manage URL query strings
 			if (config.login.enabled) {
-				scripts.push('<script src="/components/cookies-js/dist/cookies.min.js"></script>');
+				scripts.push('<script src="/bower_components/cookies-js/dist/cookies.min.js"></script>');
 				scripts.push('<script src="/login/QueryString.js"></script>');
 				scripts.push(`<script>
 					window.token = qs['key'] || Cookies.get('socketToken');
@@ -120,7 +120,7 @@ module.exports = function (pathOrHtml, resourceType, opts, cb) {
 
 			// If this bundle has sounds, inject SoundJS
 			if (opts.sound) {
-				scripts.push('<script src="/components/SoundJS/lib/soundjs-0.6.2.min.js"></script>');
+				scripts.push('<script src="/bower_components/SoundJS/lib/soundjs-0.6.2.min.js"></script>');
 			}
 
 			// Graphics must include the API script themselves before attempting to make an instance of it
