+++ package.json
@@ -26,7 +26,6 @@
     "fs-extra": "^3.0.0",
     "httpolyglot": "^0.1.1",
     "is-my-json-valid": "^2.12.3",
-    "jade": "^1.11.0",
     "json-schema-defaults": "^0.3.0",
     "minimist": "^1.2.0",
     "multer": "^1.1.0",
@@ -37,6 +36,7 @@
     "passport": "^0.3.2",
     "passport-steam": "^1.0.2",
     "passport-twitch": "^1.0.2",
+    "pug": "^2.0.0-beta11",
     "request": "^2.64.0",
     "rollbar": "^0.6.0",
     "semver": "^5.0.3",
